backend/requirements.txt
fastapi
uvicorn
python-jose
python-multipart
httpx
pydantic
